+++ src/main.js
@@ -5,4 +5,4 @@ import { translate as t, n } from '@nextcloud/l10n'
 Vue.mixin({ methods: { t, n } })
 
 const View = Vue.extend(App)
-new View().$mount('#searchengine')
+new View().$mount('#dashboardsearchbar')